+++ tests/integration/test_migrations.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.test import modify_settings, override_settings
 
 import pytest
@@ -26,11 +25,7 @@ def test_good_flow_alter_table_with_same_db_table():
 @override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
 def test_good_flow():
     # forward
-    if settings.DATABASES['default']['HOST'] in ['pg14', 'postgis14']:
-        # related to https://www.postgresql.org/message-id/3175925.1637428221%40sss.pgh.pa.us bug
-        migrate(['good_flow_app', '0045'])
-    else:
-        migrate(['good_flow_app'])
+    migrate(['good_flow_app'])
 
     # backward
     migrate(['good_flow_app', 'zero'])